+++ src/Payment.php
@@ -26,6 +26,7 @@ use ild78;
  * @method Generator list(array $terms)
  * @method self setDevice(ild78\\Device $device)
  * @method self setReturnUrl(string $https)
+ * @method self setStatus(string $status)
  */
 class Payment extends Api\AbstractObject
 {
@@ -107,7 +108,6 @@ class Payment extends Api\AbstractObject
             'type' => ild78\Sepa::class,
         ],
         'status' => [
-            'restricted' => true,
             'type' => self::STRING,
         ],
     ];
@@ -326,7 +326,7 @@ class Payment extends Api\AbstractObject
     /**
      * Refund a payment, or part of it.
      *
-     * @param integet|null $amount Amount to refund, if not present all paid amount will be refund.
+     * @param integer|null $amount Amount to refund, if not present all paid amount will be refund.
      * @return self
      * @throws ild78\Exceptions\InvalidAmountException When trying to refund more than paid.
      * @throws ild78\Exceptions\InvalidAmountException When the amount is invalid.